[angular] rework auth

+++ libs/angular/data-access/src/lib/services/auth/auth.service.ts
@@ -3,11 +3,13 @@ import { Injectable } from '@angular/core';
 import {
   AUTH_TOKEN,
   RouterExtensions,
-  StoragePlatformService
+  StoragePlatformService,
+  StoreService
 } from '@bookapp/angular/core';
 import {
   AuthPayload,
   LOGIN_MUTATION,
+  LOGOUT_MUTATION,
   ME_QUERY,
   SIGNUP_MUTATION,
   SignupCredentials,
@@ -22,7 +24,8 @@ export class AuthService {
   constructor(
     private readonly apollo: Apollo,
     private readonly storagePlatformService: StoragePlatformService,
-    private readonly routerExtensions: RouterExtensions
+    private readonly routerExtensions: RouterExtensions,
+    private readonly storeService: StoreService
   ) {}
 
   login(email: string, password: string) {
@@ -38,9 +41,10 @@ export class AuthService {
         tap(({ data }) => {
           if (data) {
             const {
-              login: { token }
+              login: { accessToken, refreshToken }
             } = data;
-            this.storagePlatformService.setItem(AUTH_TOKEN, token);
+            this.storagePlatformService.setItem(AUTH_TOKEN, refreshToken);
+            this.storeService.set(AUTH_TOKEN, accessToken);
           }
         })
       );
@@ -58,9 +62,10 @@ export class AuthService {
         tap(({ data }) => {
           if (data) {
             const {
-              signup: { token }
+              signup: { accessToken, refreshToken }
             } = data;
-            this.storagePlatformService.setItem(AUTH_TOKEN, token);
+            this.storagePlatformService.setItem(AUTH_TOKEN, refreshToken);
+            this.storeService.set(AUTH_TOKEN, accessToken);
           }
         })
       );
@@ -72,18 +77,32 @@ export class AuthService {
     });
   }
 
-  async logout() {
-    this.storagePlatformService.removeItem(AUTH_TOKEN);
-    this.routerExtensions.navigate(['auth'], {
-      // for nativescript
-      clearHistory: true,
-      transition: {
-        name: 'flip',
-        duration: 300,
-        curve: 'linear'
-      }
-    });
+  logout() {
+    return this.apollo
+      .mutate<{ logout: boolean }>({
+        mutation: LOGOUT_MUTATION,
+        variables: {
+          refreshToken: this.storagePlatformService.getItem(AUTH_TOKEN)
+        }
+      })
+      .pipe(
+        tap(async ({ data }) => {
+          if (data.logout) {
+            this.storagePlatformService.removeItem(AUTH_TOKEN);
+            this.storeService.remove(AUTH_TOKEN);
+            this.routerExtensions.navigate(['auth'], {
+              // for nativescript
+              clearHistory: true,
+              transition: {
+                name: 'flip',
+                duration: 300,
+                curve: 'linear'
+              }
+            });
 
-    await this.apollo.getClient().clearStore();
+            await this.apollo.getClient().clearStore();
+          }
+        })
+      );
   }
 }
